requirements.txt
async-timeout==4.0.3
kafka-python==2.0.2
mypy==0.991
mypy-extensions==1.0.0
numpy==1.24.4
pandas==2.0.3
ping3==4.0.4
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.0
six==1.16.0
tomli==2.0.1
types-redis==4.3.21.4
typing_extensions==4.8.0
tzdata==2023.3
